+++ lib/haml_lint/linter/unnecessary_string_output.rb
@@ -31,9 +31,16 @@ module HamlLint
 
     def outputs_string_literal?(script_node)
       tree = parse_ruby(script_node.script)
-      [:str, :dstr].include?(tree.type)
+      [:str, :dstr].include?(tree.type) &&
+        starts_with_reserved_character?(tree.children.first)
     rescue ::Parser::SyntaxError # rubocop:disable Lint/HandleExceptions
       # Gracefully ignore syntax errors, as that's managed by a different linter
+    end
+
+    # Returns whether a string starts with a character that would otherwise be
+    # given special treatment, thus making enclosing it in a string necessary.
+    def starts_with_reserved_character?(string)
+      string !~ %r{\A\s*[/#-=%~]}
     end
   end
 end