+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## master (unreleased)
 
 * Change required Ruby version from 1.9.3+ to 2.0.0+ since 1.9.3 has been EOLed
+* Fix false positives in `UnnecessaryStringOutput` for strings starting with
+  reserved HAML characters (where enclosing in a string is required)
 
 ## 0.13.0